+++ assets/scripts/routes/atelier-pages.js
@@ -8,7 +8,7 @@ import gitAgent from '../gitAgent'
 import { replaceComponent } from '../routeComponentLifeCycle'
 import store from '../store'
 import { handleErrors, logMessage, makeFileNameFromTitle } from '../utils'
-import { setCurrentRepositoryFromQuerystring } from '../actions'
+import { setCurrentRepositoryFromQuerystring } from '../actions/current-repository.js'
 import PageContenu from '../components/screens/PageContenu.svelte'
 import { deletePage, createPage, updatePage } from './../actions/page'
 import { makeAtelierListPageURL } from './urls.js'
