+++ assets/scripts/routes/settings.js
@@ -5,11 +5,11 @@ import gitAgent from '../gitAgent'
 import { replaceComponent } from '../routeComponentLifeCycle'
 import store from '../store'
 import {
+  deleteRepository,
   getCurrentRepoPages,
   getCurrentRepoArticles,
   setCurrentRepositoryFromQuerystring,
-} from '../actions'
-import { deleteRepository } from '../actions/repository.js'
+} from '../actions/current-repository.js'
 import { handleErrors } from '../utils'
 import Settings from '../components/screens/Settings.svelte'
 import page from 'page'
@@ -49,7 +49,7 @@ function mapStateToProps(state) {
   const blogFile = state.pages && state.pages.find(p => p.path === 'blog.md')
   const currentRepository = store.state.currentRepository
 
-  if(!currentRepository){
+  if (!currentRepository) {
     throw new TypeError('currentRepository is undefined')
   }
 
@@ -72,7 +72,7 @@ export default ({ querystring }) => {
 
   const currentRepository = store.state.currentRepository
 
-  if(!currentRepository){
+  if (!currentRepository) {
     throw new TypeError('currentRepository is undefined')
   }
 
@@ -92,10 +92,7 @@ export default ({ querystring }) => {
 
   settings.$on('update-theme', ({ detail: { theme } }) => {
     gitAgent
-      .writeCustomCSS(
-        currentRepository,
-        theme.css
-      )
+      .writeCustomCSS(currentRepository, theme.css)
       .then(_ => {
         store.mutations.setTheme(store.state.theme.css)
         store.state.buildStatus.setBuildingAndCheckStatusLater(10000)
@@ -122,10 +119,7 @@ export default ({ querystring }) => {
 
   if (!store.state.theme.css) {
     gitAgent
-      .getFile(
-        currentRepository,
-        gitAgent.customCSSPath
-      )
+      .getFile(currentRepository, gitAgent.customCSSPath)
       .then(content => {
         store.mutations.setTheme(content)
       })
